Thread struct and error handling, signals
A thread struct is now passed to each thread instead of just the socket reference. This is intended to be used for communication with threads.
Improved error handling through signals. The return of each read is checked, and a relevant error value is returned from the function which
is used for communication. A termination message of the form !"£$term$£"! causes an interrupt to be triggered and the thread, in the case of the server, or the client will exit, freeing memory as it does so.